requirements.txt
requests>=2.28.0
newspaper3k>=0.2.8
transformers>=4.21.0
torch>=1.12.0
python-dotenv>=0.19.0
ollama>=0.1.0
